+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4054,7 +4054,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KY7C3C0MBQRJ6JBF7QRWC2E7"),
+  id: z.string().default("01KY9CSMPW6GVZH22V2ZZ1RAA9"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),